+++ Dockerfiles/pgadmin4.Dockerfile
@@ -1,1 +1,1 @@
-FROM dpage/pgadmin4:6.16
+FROM dpage/pgadmin4:6.17
